// src/models/components/updatevisibilityrequest.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v4-mini";
import {
  Visibility,
  Visibility$outboundSchema,
} from "./visibility.js";

export type UpdateVisibilityRequest = {
  /**
   * Image visibility setting
   */
  visibility: Visibility;
};

/** @internal */
export type UpdateVisibilityRequest$Outbound = {
  visibility: string;
};

/** @internal */
export const UpdateVisibilityRequest$outboundSchema: z.ZodMiniType<
  UpdateVisibilityRequest$Outbound,
  UpdateVisibilityRequest
> = z.object({
  visibility: Visibility$outboundSchema,
});

export function updateVisibilityRequestToJSON(
  updateVisibilityRequest: UpdateVisibilityRequest,
): string {
  return JSON.stringify(
    UpdateVisibilityRequest$outboundSchema.parse(updateVisibilityRequest),
  );
}